#1 -added CM-matrices for comparison - shifted plot function - first experiment results look appropriate
