+++ pre_investigations/python/solver_investigations/CM_matrices/CM_nodes2.json
@@ -1,0 +1,1 @@
+[[[0.0, 1.0, 2.0, 3.0], [-1.0, 0.0, 4.0, 5.0], [-2.0, -4.0, 0.0, 6.0], [-3.0, -5.0, -6.0, 0.0]], [[0.0, 1.0, 6.0, 2.0], [-1.0, 0.0, 3.0, 4.0], [-6.0, -3.0, 0.0, 5.0], [-2.0, -4.0, -5.0, 0.0]], [[0.0, 1.0, 2.0, 3.0], [-1.0, 0.0, 6.0, 4.0], [-2.0, -6.0, 0.0, 5.0], [-3.0, -4.0, -5.0, 0.0]], [[0.0, 4.0, 1.0, 2.0], [-4.0, 0.0, 6.0, 5.0], [-1.0, -6.0, 0.0, 3.0], [-2.0, -5.0, -3.0, 0.0]], [[0.0, 5.0, 1.0, 6.0], [-5.0, 0.0, 2.0, 3.0], [-1.0, -2.0, 0.0, 4.0], [-6.0, -3.0, -4.0, 0.0]], [[0.0, 1.0, 2.0, 3.0], [-1.0, 0.0, 4.0, 5.0], [-2.0, -4.0, 0.0, 6.0], [-3.0, -5.0, -6.0, 0.0]], [[0.0, 5.0, 1.0, 2.0], [-5.0, 0.0, 6.0, 3.0], [-1.0, -6.0, 0.0, 4.0], [-2.0, -3.0, -4.0, 0.0]], [[0.0, 1.0, 5.0, 2.0], [-1.0, 0.0, 3.0, 6.0], [-5.0, -3.0, 0.0, 4.0], [-2.0, -6.0, -4.0, 0.0]], [[0.0, 6.0, 1.0, 2.0], [-6.0, 0.0, 3.0, 4.0], [-1.0, -3.0, 0.0, 5.0], [-2.0, -4.0, -5.0, 0.0]], [[0.0, 6.0, 1.0, 2.0], [-6.0, 0.0, 3.0, 4.0], [-1.0, -3.0, 0.0, 5.0], [-2.0, -4.0, -5.0, 0.0]], [[0.0, 4.0, 1.0, 2.0], [-4.0, 0.0, 6.0, 5.0], [-1.0, -6.0, 0.0, 3.0], [-2.0, -5.0, -3.0, 0.0]], [[0.0, 5.0, 1.0, 2.0], [-5.0, 0.0, 6.0, 3.0], [-1.0, -6.0, 0.0, 4.0], [-2.0, -3.0, -4.0, 0.0]], [[0.0, 1.0, 2.0, 3.0], [-1.0, 0.0, 4.0, 5.0], [-2.0, -4.0, 0.0, 6.0], [-3.0, -5.0, -6.0, 0.0]], [[0.0, 4.0, 1.0, 2.0], [-4.0, 0.0, 5.0, 3.0], [-1.0, -5.0, 0.0, 6.0], [-2.0, -3.0, -6.0, 0.0]], [[0.0, 5.0, 1.0, 2.0], [-5.0, 0.0, 6.0, 3.0], [-1.0, -6.0, 0.0, 4.0], [-2.0, -3.0, -4.0, 0.0]], [[0.0, 1.0, 6.0, 2.0], [-1.0, 0.0, 3.0, 4.0], [-6.0, -3.0, 0.0, 5.0], [-2.0, -4.0, -5.0, 0.0]], [[0.0, 5.0, 1.0, 2.0], [-5.0, 0.0, 3.0, 4.0], [-1.0, -3.0, 0.0, 6.0], [-2.0, -4.0, -6.0, 0.0]], [[0.0, 5.0, 4.0, 1.0], [-5.0, 0.0, 2.0, 3.0], [-4.0, -2.0, 0.0, 6.0], [-1.0, -3.0, -6.0, 0.0]], [[0.0, 5.0, 4.0, 1.0], [-5.0, 0.0, 2.0, 6.0], [-4.0, -2.0, 0.0, 3.0], [-1.0, -6.0, -3.0, 0.0]], [[0.0, 4.0, 5.0, 1.0], [-4.0, 0.0, 2.0, 6.0], [-5.0, -2.0, 0.0, 3.0], [-1.0, -6.0, -3.0, 0.0]], [[0.0, 1.0, 2.0, 3.0], [-1.0, 0.0, 4.0, 5.0], [-2.0, -4.0, 0.0, 6.0], [-3.0, -5.0, -6.0, 0.0]], [[0.0, 5.0, 1.0, 2.0], [-5.0, 0.0, 6.0, 3.0], [-1.0, -6.0, 0.0, 4.0], [-2.0, -3.0, -4.0, 0.0]], [[0.0, 5.0, 1.0, 2.0], [-5.0, 0.0, 3.0, 4.0], [-1.0, -3.0, 0.0, 6.0], [-2.0, -4.0, -6.0, 0.0]], [[0.0, 5.0, 1.0, 2.0], [-5.0, 0.0, 3.0, 6.0], [-1.0, -3.0, 0.0, 4.0], [-2.0, -6.0, -4.0, 0.0]], [[0.0, 4.0, 1.0, 2.0], [-4.0, 0.0, 5.0, 6.0], [-1.0, -5.0, 0.0, 3.0], [-2.0, -6.0, -3.0, 0.0]], [[0.0, 6.0, 1.0, 2.0], [-6.0, 0.0, 3.0, 4.0], [-1.0, -3.0, 0.0, 5.0], [-2.0, -4.0, -5.0, 0.0]], [[0.0, 5.0, 1.0, 2.0], [-5.0, 0.0, 6.0, 3.0], [-1.0, -6.0, 0.0, 4.0], [-2.0, -3.0, -4.0, 0.0]], [[0.0, 6.0, 1.0, 2.0], [-6.0, 0.0, 3.0, 4.0], [-1.0, -3.0, 0.0, 5.0], [-2.0, -4.0, -5.0, 0.0]], [[0.0, 6.0, 1.0, 2.0], [-6.0, 0.0, 3.0, 4.0], [-1.0, -3.0, 0.0, 5.0], [-2.0, -4.0, -5.0, 0.0]], [[0.0, 6.0, 5.0, 4.0], [-6.0, 0.0, 1.0, 2.0], [-5.0, -1.0, 0.0, 3.0], [-4.0, -2.0, -3.0, 0.0]]]
